Adjust flow to get route snapper
